transferred data to another component
